refactor(widgets): Widgets separated in new file
